+++ doc/Changelog.md
@@ -1,3 +1,4 @@
+* text typed in the search field is now highlighted during typing
 * bug fixes
 
 ### version v5.3.4 (July 15, 2021)